notebooks: Merge focus and switch-page handling of notebooks
The switch-page handler in callbacks.c handled a lot of GUI work needed
when document tabs are changed. The same work is needed when the focus between
notebooks changes (this is the same as a tab switch but does not cause the
signal).
